networkrail_consumer/pyproject.toml
[tool.poetry]
name = "networkrail-consumer"
version = "0.1.0"
description = ""
authors = ["zkan <zkan.cs@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
stomp-py = "^8.1.2"
pytz = "^2024.1"
kafka-python-ng = "2.2.2"
google-cloud-storage = "^2.18.2"
grpcio = "^1.56.0 "

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
